Added model and migrations for eligibilities (#232)

+++ lib/tasks/populate.rake
@@ -11,6 +11,7 @@ namespace :db do
     Rails.application.eager_load! # Load all models
     ActiveRecord::Base.transaction do
       ActiveRecord::Base.descendants.each do |model|
+        next if model.to_s == 'ApplicationRecord'
         next if model.to_s.starts_with? 'ActiveRecord::'
         model.destroy_all
         ActiveRecord::Base.connection.reset_pk_sequence!(model.table_name)
